backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
pymysql==1.1.0
redis==5.0.1

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 数据采集
requests==2.31.0
aiohttp==3.9.1
bilibili-api-python==17.4.1

# 数据分析
pandas==2.1.3
numpy==1.26.2

# NLP
jieba==0.42.1
snownlp==0.12.3

# 定时任务
apscheduler==3.10.4

# 数据导出
openpyxl==3.1.2

# 工具
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
